Pull up logic from GitTabController to RootController
This makes more sense architecturally since viewing the (un)staged
changes for files is not a concern of the git tab
