+++ lib/controllers/git-tab-controller.js
@@ -3,7 +3,7 @@
 
 import path from 'path';
 import etch from 'etch';
-import {CompositeDisposable, Disposable} from 'event-kit';
+import {Disposable} from 'event-kit';
 
 import GitTabView from '../views/git-tab-view';
 import ModelObserver from '../models/model-observer';
@@ -33,14 +33,7 @@ export default class GitTabController {
     etch.initialize(this);
 
     this.element.addEventListener('focusin', this.rememberLastFocus);
-    this.subscriptions = new CompositeDisposable();
-    this.subscriptions.add(
-      new Disposable(() => this.element.removeEventListener('focusin', this.rememberLastFocus)),
-      this.props.commandRegistry.add('atom-workspace', {
-        'github:view-unstaged-changes-for-current-file': this.viewUnstagedChangesForCurrentFile,
-        'github:view-staged-changes-for-current-file': this.viewStagedChangesForCurrentFile,
-      }),
-    );
+    this.subscriptions = new Disposable(() => this.element.removeEventListener('focusin', this.rememberLastFocus));
   }
 
   serialize() {
@@ -267,41 +260,6 @@ export default class GitTabController {
       return repository.stageFilesFromParentCommit(filePaths);
     } else {
       return repository.unstageFiles(filePaths);
-    }
-  }
-
-  @autobind
-  viewUnstagedChangesForCurrentFile() {
-    const editor = this.props.workspace.getActiveTextEditor();
-    const absFilePath = editor.getPath();
-    const repoPath = this.getActiveRepository().getWorkingDirectoryPath();
-    if (absFilePath.startsWith(repoPath)) {
-      const filePath = absFilePath.slice(repoPath.length + 1);
-      this.quietlySelectItem(filePath, 'unstaged');
-      return this.props.showFilePatchForPath(filePath, 'unstaged', {
-        activate: true,
-        lineNumber: editor.getCursorBufferPosition().row + 1,
-      });
-    } else {
-      throw new Error(`${absFilePath} does not belong to repo ${repoPath}`);
-    }
-  }
-
-  @autobind
-  viewStagedChangesForCurrentFile() {
-    const editor = this.props.workspace.getActiveTextEditor();
-    const absFilePath = editor.getPath();
-    const repoPath = this.getActiveRepository().getWorkingDirectoryPath();
-    if (absFilePath.startsWith(repoPath)) {
-      const filePath = absFilePath.slice(repoPath.length + 1);
-      this.quietlySelectItem(filePath, 'staged');
-      return this.props.showFilePatchForPath(filePath, 'staged', {
-        activate: true,
-        lineNumber: editor.getCursorBufferPosition().row + 1,
-        amending: this.props.isAmending,
-      });
-    } else {
-      throw new Error(`${absFilePath} does not belong to repo ${repoPath}`);
     }
   }
