+++ lib/controllers/root-controller.js
@@ -146,6 +146,8 @@ export default class RootController extends React.Component {
           <Command command="github:toggle-github-tab" callback={this.githubTabTracker.toggle} />
           <Command command="github:toggle-github-tab-focus" callback={this.githubTabTracker.toggleFocus} />
           <Command command="github:clone" callback={this.openCloneDialog} />
+          <Command command="github:view-unstaged-changes-for-current-file" callback={this.viewUnstagedChangesForCurrentFile} />
+          <Command command="github:view-staged-changes-for-current-file" callback={this.viewStagedChangesForCurrentFile} />
         </Commands>
         {this.renderStatusBarTile()}
         {this.renderPanels()}
@@ -579,6 +581,40 @@ export default class RootController extends React.Component {
     viewElement.focus();
   }
 
+  @autobind
+  viewUnstagedChangesForCurrentFile() {
+    const editor = this.props.workspace.getActiveTextEditor();
+    const absFilePath = editor.getPath();
+    const repoPath = this.props.repository.getWorkingDirectoryPath();
+    if (absFilePath.startsWith(repoPath)) {
+      const filePath = absFilePath.slice(repoPath.length + 1);
+      this.quietlySelectItem(filePath, 'unstaged');
+      return this.showFilePatchForPath(filePath, 'unstaged', {
+        activate: true,
+        lineNumber: editor.getCursorBufferPosition().row + 1,
+      });
+    } else {
+      throw new Error(`${absFilePath} does not belong to repo ${repoPath}`);
+    }
+  }
+
+  @autobind
+  viewStagedChangesForCurrentFile() {
+    const editor = this.props.workspace.getActiveTextEditor();
+    const absFilePath = editor.getPath();
+    const repoPath = this.props.repository.getWorkingDirectoryPath();
+    if (absFilePath.startsWith(repoPath)) {
+      const filePath = absFilePath.slice(repoPath.length + 1);
+      this.quietlySelectItem(filePath, 'staged');
+      return this.showFilePatchForPath(filePath, 'staged', {
+        activate: true,
+        lineNumber: editor.getCursorBufferPosition().row + 1,
+        amending: this.state.isAmending,
+      });
+    } else {
+      throw new Error(`${absFilePath} does not belong to repo ${repoPath}`);
+    }
+  }
 
   @autobind
   goToDiffLine(lineNumber) {
